pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kbskills"
version = "0.1.0"
description = "Local knowledge base system with GraphRAG and Skills-driven outline generation"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "lightrag-hku>=1.0",
    "google-genai>=1.0",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "python-pptx>=0.6",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "html2text>=2024.2",
    "youtube-transcript-api>=0.6",
    "yt-dlp>=2024.0",
    "rich>=13.7",
    "pyyaml>=6.0",
    "numpy>=1.26",
]

[project.scripts]
kbskills = "kbskills.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/kbskills"]
